requirements.txt
crewai
load_dotenv
langchain-community        # Extends LangChain with community-built tools
langchain-core             # Core components for building LLM apps
streamlit                  # Creates user-friendly web interfaces 
langchain                  # Enables LLM-based workflows and integrations
python-dotenv              # Manages environment variables securely
langchain_groq             # Integrates GROQ features with LangChainS
fastapi
uvicorn
crewai-tools
langchain-huggingface